udpate

+++ src/ambient/contracts/libraries/KnockoutLiq.sol
@@ -7,7 +7,6 @@ pragma solidity 0.8.19;
  *         tick pivot. In that sense knockout liquidity behaves like a "non-reversible
  *         limit order" seen in the traditional limit order book. */
 library KnockoutLiq {
-
     /* @notice Defines a currently active knockout liquidity bump point that exists on
      *         a specific AMM curve at a specific tick/direction.
      *
@@ -75,37 +74,35 @@ library KnockoutLiq {
     }
 
     /* @notice Resets all fields on a existing pivot struct. */
-    function deletePivot (KnockoutPivot storage pivot) internal {
+    function deletePivot(KnockoutPivot storage pivot) internal {
         pivot.lots_ = 0;
         pivot.pivotTime_ = 0;
         pivot.rangeTicks_ = 0;
     }
-    
+
     /* @notice Encodes a hash key for a given knockout pivot point.
      * @param pool The hash index of the AMM pool.
      * @param isBid If true indicates the knockout pivot is on the bid side.
      * @param tick The tick index of the knockout pivot.
      * @return Unique hash key mapping to the pivot struct. */
-    function encodePivotKey (bytes32 pool, bool isBid, int24 tick)
-        internal pure returns (bytes32) {
+    function encodePivotKey(bytes32 pool, bool isBid, int24 tick) internal pure returns (bytes32) {
         return keccak256(abi.encode(pool, isBid, tick));
     }
 
     /* @notice Encodes a hash key for a knockout pivot given a pos location struct. */
-    function encodePivotKey (KnockoutPosLoc memory loc, bytes32 pool)
-        internal pure returns (bytes32) {
+    function encodePivotKey(KnockoutPosLoc memory loc, bytes32 pool) internal pure returns (bytes32) {
         return encodePivotKey(pool, loc.isBid_, knockoutTick(loc));
     }
 
     /* @notice Determines which tick side is the pivot point based on whether the pivot
      *         is a bid or ask. */
-    function knockoutTick (KnockoutPosLoc memory loc) internal pure returns (int24) {
+    function knockoutTick(KnockoutPosLoc memory loc) internal pure returns (int24) {
         return loc.isBid_ ? loc.lowerTick_ : loc.upperTick_;
     }
 
-    function tickRange (KnockoutPosLoc memory loc) internal pure returns (uint16) {
+    function tickRange(KnockoutPosLoc memory loc) internal pure returns (uint16) {
         uint24 range = uint24(loc.upperTick_ - loc.lowerTick_);
-        require (range < type(uint16).max);
+        require(range < type(uint16).max);
         return uint16(range);
     }
 
@@ -115,11 +112,12 @@ library KnockoutLiq {
      * @param owner The claimant of the liquidity position.
      * @param pivotTime The timestamp of when the pivot tranche was created
      * @return Unique hash key to position. */
-    function encodePosKey (KnockoutPosLoc memory loc,
-                           bytes32 pool, address owner, uint32 pivotTime)
-        internal pure returns (bytes32) {
-        return keccak256(abi.encode(pool, owner, loc.isBid_,
-                                    loc.lowerTick_, loc.upperTick_, pivotTime));
+    function encodePosKey(KnockoutPosLoc memory loc, bytes32 pool, address owner, uint32 pivotTime)
+        internal
+        pure
+        returns (bytes32)
+    {
+        return keccak256(abi.encode(pool, owner, loc.isBid_, loc.lowerTick_, loc.upperTick_, pivotTime));
     }
     /* @notice Commits a now-crossed Knockout pivot to the merkle history for that tick
      *         location.
@@ -127,8 +125,8 @@ library KnockoutLiq {
      * @param pivot The most recent pivot state. Should not call this unless the pivot has
      *              just been knocked out.
      * @param feeMileage The in-range fee mileage at the time of knockout crossing. */
-    function commitKnockout (KnockoutMerkle storage merkle,
-                             KnockoutPivot memory pivot, uint64 feeMileage) internal {
+
+    function commitKnockout(KnockoutMerkle storage merkle, KnockoutPivot memory pivot, uint64 feeMileage) internal {
         merkle.merkleRoot_ = rootLink(merkle, commitEntropySalt());
         merkle.pivotTime_ = pivot.pivotTime_;
         merkle.feeMileage_ = feeMileage;
@@ -152,39 +150,32 @@ library KnockoutLiq {
      *      and raises the cost of a long-range collision attack to 2^140 hashes, which 
      *      is outright infeasible. */
     function commitEntropySalt() internal view returns (uint160) {
-        return uint160(uint256(blockhash(block.number-1)));
+        return uint160(uint256(blockhash(block.number - 1)));
     }
 
     /* @notice Converts the most recent Merkle state to a 160-bit Merkle root hash. */
-    function rootLink (KnockoutMerkle memory merkle, uint160 salt)
-        private pure returns (uint160) {
+    function rootLink(KnockoutMerkle memory merkle, uint160 salt) private pure returns (uint160) {
         return rootLink(merkle.merkleRoot_, merkle.pivotTime_, merkle.feeMileage_, salt);
     }
 
     /* @notice Converts the most current Merkle state params to 160-bit Merkle hash.*/
-    function rootLink (uint160 root, uint32 pivotTime, uint64 feeMileage, uint160 salt)
-        private pure returns (uint160) {
+    function rootLink(uint160 root, uint32 pivotTime, uint64 feeMileage, uint160 salt) private pure returns (uint160) {
         return rootLink(root, encodeChainLink(pivotTime, feeMileage, salt));
     }
 
     /* @notice Hashes together the previous Merkle root with the encoded chain step. */
-    function rootLink (uint160 root, uint256 chainLink)
-        private pure returns (uint160) {
+    function rootLink(uint160 root, uint256 chainLink) private pure returns (uint160) {
         bytes32 hash = keccak256(abi.encode(root, chainLink));
         return uint160(uint256(hash) >> 96);
     }
 
     /* @notice Tightly packs the 32-bit pivot time with the 64-bit fee mileage and the salt. */
-    function encodeChainLink (uint32 pivotTime, uint64 feeMileage, uint160 salt)
-        private pure returns (uint256)  {
-        return (uint256(salt) << 96) +
-            (uint256(pivotTime) << 64) +
-            uint256(feeMileage);
+    function encodeChainLink(uint32 pivotTime, uint64 feeMileage, uint160 salt) private pure returns (uint256) {
+        return (uint256(salt) << 96) + (uint256(pivotTime) << 64) + uint256(feeMileage);
     }
 
     /* @notice Decodes a tightly packed chain link into the pivot time and fee mileage */
-    function decodeChainLink (uint256 entry)
-        private pure returns (uint32 pivotTime, uint64 feeMileage)  {
+    function decodeChainLink(uint256 entry) private pure returns (uint32 pivotTime, uint64 feeMileage) {
         pivotTime = uint32((entry << 160) >> 224);
         feeMileage = uint64((entry << 192) >> 192);
     }
@@ -199,24 +190,26 @@ library KnockoutLiq {
      *
      * @return The 32-bit Knockout tranche pivot time and 64-bit fee mileage at the start of
      *         the proof. */
-    function proveHistory (KnockoutMerkle memory merkle, uint160 proofRoot,
-                           uint256[] memory proof)
-        internal pure returns (uint32, uint64) {
+    function proveHistory(KnockoutMerkle memory merkle, uint160 proofRoot, uint256[] memory proof)
+        internal
+        pure
+        returns (uint32, uint64)
+    {
         // If we're only looking at the most recent knockout, it's still stored raw
         // and doesn't need a proof.
-        return proof.length == 0 ?
-            (merkle.pivotTime_, merkle.feeMileage_) :
-            proveSteps(merkle, proofRoot, proof);
+        return proof.length == 0 ? (merkle.pivotTime_, merkle.feeMileage_) : proveSteps(merkle, proofRoot, proof);
     }
 
     /* @notice Verifies a non-empty Merkle proof. */
-    function proveSteps (KnockoutMerkle memory merkle, uint160 proofRoot,
-                         uint256[] memory proof)
-        private pure returns (uint32, uint64) {
+    function proveSteps(KnockoutMerkle memory merkle, uint160 proofRoot, uint256[] memory proof)
+        private
+        pure
+        returns (uint32, uint64)
+    {
         uint160 incrRoot = proofRoot;
         unchecked {
             // Iterate by 1 loop will run out of gas far before overflowing 256 bits
-            for (uint i = 0; i < proof.length; ++i) {
+            for (uint256 i = 0; i < proof.length; ++i) {
                 incrRoot = rootLink(incrRoot, proof[i]);
             }
         }
@@ -224,7 +217,6 @@ library KnockoutLiq {
         require(incrRoot == merkle.merkleRoot_, "KP");
         return decodeChainLink(proof[0]);
     }
-
 
     /* @notice Verifies that a given knockout location is valid relative to the curve
      *         price and the pool's current knockout parameters. If not, the call will
@@ -253,19 +245,15 @@ library KnockoutLiq {
      *                             be placed on a tick index that's a multiple of the width. 
      *                             Can be used to restrict density of knockout orders, beyond 
      *                             the normal pool tick size. */
-    function assertValidPos (KnockoutPosLoc memory loc, int24 priceTick, 
-                             uint8 knockoutBits) internal pure {
-        (bool enabled, uint8 width, bool inside, bool onGrid) =
-            unpackBits(knockoutBits);
+    function assertValidPos(KnockoutPosLoc memory loc, int24 priceTick, uint8 knockoutBits) internal pure {
+        (bool enabled, uint8 width, bool inside, bool onGrid) = unpackBits(knockoutBits);
 
-        require(enabled && gridOkay(loc, width, onGrid) &&
-                spreadOkay(loc, priceTick, inside), "KV");
+        require(enabled && gridOkay(loc, width, onGrid) && spreadOkay(loc, priceTick, inside), "KV");
     }
 
     /* @notice Evaluates whether the placement and width of a knockout pivot candidate
      *         conforms to the grid parameters. */
-    function gridOkay (KnockoutPosLoc memory loc, uint8 widthBits, bool mustBeOnGrid)
-        private pure returns (bool) {
+    function gridOkay(KnockoutPosLoc memory loc, uint8 widthBits, bool mustBeOnGrid) private pure returns (bool) {
         uint24 width = uint24(loc.upperTick_ - loc.lowerTick_);
         bool rightWidth = width == uint24(1) << widthBits;
 
@@ -278,11 +266,10 @@ library KnockoutLiq {
 
     /* @notice Evaluates whether the placement of a knockout pivot candidates conforms
      *         to the parameters relative to the curve's current price tick. */
-    function spreadOkay (KnockoutPosLoc memory loc, int24 priceTick,
-                         bool inside) internal pure returns (bool) {
+    function spreadOkay(KnockoutPosLoc memory loc, int24 priceTick, bool inside) internal pure returns (bool) {
         // Checks to see whether the range order is placed directionally correct relative
         // to the current tick price. If inside is true, then the range order can be placed
-        // with the curve price inside the range. 
+        // with the curve price inside the range.
         // Otherwise bids must have the entire range below the curve price, and asks must
         // have the entire range above the curve price.
         if (loc.isBid_) {
@@ -303,8 +290,11 @@ library KnockoutLiq {
      *                 price.
      * @return onGrid True if new knockout range orders are restricted to ticks that
      *                are multiples of the width size. */
-    function unpackBits (uint8 knockoutBits) private pure returns
-        (bool enabled, uint8 widthBits, bool inside, bool onGrid) {
+    function unpackBits(uint8 knockoutBits)
+        private
+        pure
+        returns (bool enabled, uint8 widthBits, bool inside, bool onGrid)
+    {
         widthBits = uint8(knockoutBits & 0x0F);
         uint8 flagBits = uint8(knockoutBits & 0x30) >> 4;
